+++ Apps/Database/src/Models/MessagingVerification.cs
@@ -90,7 +90,7 @@ namespace HealthGateway.Database.Models
         /// Gets or sets the attemped verification count.
         /// </summary>
         [Required]
-        public byte VerificationAttempts { get; set; }
+        public int VerificationAttempts { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether the messageing verification was deleted.
